ssh auto complete based on .ssh/config and known_hosts
All entries in ssh client config excluding wildcard entries as well
as ssh known_hosts whilst parsing enties with no rev DNS (brackets are
interpreted) are added to ssh, scp and deploy_bgrc auto complete.
Hashed known hosts are skipped of course

Removed commented-out code for ssh auto complete base on history. -
This is slow if the history is large and it has not been used for a
while.